chore: remove ember implementations from `migration-graph` (#436)
