+++ packages/migration-graph-ember/src/entities/ember-app-package.ts
@@ -1,7 +1,8 @@
 import { relative, resolve } from 'path';
-import { type PackageOptions, Package } from '@rehearsal/migration-graph-shared';
+import { type PackageOptions, Package, Graph, ModuleNode } from '@rehearsal/migration-graph-shared';
 
 import { InternalState } from './InternalState';
+import { EmberAppPackageGraph } from './ember-app-package-graph';
 
 export type EmberPackageOptions = PackageOptions;
 
@@ -15,7 +16,15 @@ export class EmberAppPackage extends Package {
 
   get excludePatterns(): Array<string> {
     // TODO Determine ember-config directory from package.json entry
-    return ['dist', 'config', 'ember-config', 'tests', '@ember/*', 'public'];
+    const files = [
+      '.ember-cli.js',
+      'ember-cli-build.js',
+      'ember-config.js',
+      'index.js',
+      'testem.js',
+    ];
+    const directories = ['dist', 'config', 'ember-config', 'tests', '@ember/*', 'public'];
+    return [...directories, ...files];
   }
 
   get includePatterns(): Array<string> {
@@ -86,5 +95,15 @@ export class EmberAppPackage extends Package {
     );
 
     return this;
+  }
+
+  createModuleGraph(options = {}): Graph<ModuleNode> {
+    if (this.files) {
+      return this.files;
+    }
+
+    this.files = new EmberAppPackageGraph(this, options).discover();
+
+    return this.files;
   }
 }
