+++ packages/migration-graph-ember/src/index.ts
@@ -19,4 +19,17 @@ export {
   getEmberAddonName,
 } from './utils/ember';
 
+export { discoverEmberPackages } from './utils/discover-ember-packages';
 export { discoverServiceDependencies } from './utils/discover-ember-service-dependencies';
+
+export {
+  EmberAppPackageGraph,
+  EmberAppPackageGraphOptions,
+} from './entities/ember-app-package-graph';
+export {
+  EmberAddonPackageGraph,
+  EmberAddonPackageGraphOptions,
+} from './entities/ember-addon-package-graph';
+
+export { EmberAppProjectGraph } from './entities/ember-app-project-graph';
+export { EmberAddonProjectGraph } from './entities/ember-addon-project-graph';
