+++ packages/migration-graph-ember/tests/entities/ember-addon-package-graph.test.ts
@@ -1,0 +1,21 @@
+import { describe, expect, test } from 'vitest';
+import { getEmberProject, setupProject } from '@rehearsal/test-support';
+import { EmberAddonPackage } from '../../src/entities/ember-addon-package';
+import { EmberAddonPackageGraph } from '../../src/entities/ember-addon-package-graph';
+
+describe('EmberAddonPackageDependencyGraph', () => {
+  test('should create an edge between app/components/<file>.js and addon/components/<file>.js', async () => {
+    const project = getEmberProject('addon');
+
+    await setupProject(project);
+
+    const addonPackage = new EmberAddonPackage(project.baseDir);
+
+    const addonPackageGraph = new EmberAddonPackageGraph(addonPackage);
+    addonPackageGraph.discover();
+
+    const implNode = addonPackageGraph.getGraph().getNode('addon/components/greet.js');
+    const interfaceNode = addonPackageGraph.getGraph().getNode('app/components/greet.js');
+    expect(interfaceNode.adjacent.has(implNode)).toBe(true);
+  });
+});
