+++ packages/migration-graph-ember/tests/utils/discover-ember-packages.test.ts
@@ -1,0 +1,8 @@
+import { describe, expect, test } from 'vitest';
+
+// TODO: should have a test
+describe('discover-ember-packages', () => {
+  test.todo('should have test', () => {
+    expect(true).toBe(false);
+  });
+});
