+++ packages/migration-graph-shared/tests/tsconfig.json
@@ -4,5 +4,5 @@
     "rootDir": "../",
     "outDir": "../dist/test/"
   },
-  "include": [".", "../src"]
+  "include": [".", "../src", "../../migration-graph-ember/src/create-package-depdendency-graph.ts"]
 }
