+++ packages/migration-graph/package.json
@@ -21,7 +21,6 @@
     "@rehearsal/migration-graph-shared": "workspace:*",
     "@swc/core": "^1.3.2",
     "debug": "^4.3.4",
-    "dependency-cruiser": "^11.15.0",
     "enhanced-resolve": "^5.10.0",
     "fast-glob": "^3.2.12",
     "fs-extra": "^10.1.0",
