+++ packages/migration-graph/src/index.ts
@@ -1,14 +1,5 @@
-export { buildMigrationGraph, discoverEmberPackages } from './migration-graph';
-export { GraphNode } from './utils/graph-node';
-export { Graph } from './utils/graph';
-
-export {
-  isModuleRelative,
-  isModuleNonRelative,
-  isDirectoryPackage,
-  getMainEntrypoint,
-  resolveRelativeModule,
-} from './module-resolution';
+// This is consumed by CLI. We can rename the import and drop this.
+export { discoverEmberPackages } from '@rehearsal/migration-graph-ember';
 
 export {
   getMigrationStrategy,
@@ -16,5 +7,3 @@ export {
   type MigrationStrategyOptions,
   type SourceFile,
 } from './migration-strategy';
-
-export type { UniqueGraphNode, PackageNode, ModuleNode } from './types';
