+++ packages/migration-graph/tests/migration-strategy.test.ts
@@ -4,10 +4,8 @@ import {
   getLibraryWithEntrypoint,
   getEmberProjectFixture,
 } from '@rehearsal/test-support';
-import { DetectedSource } from '../src/migration-graph';
+import { SourceType } from '../src/source-type';
 import { getMigrationStrategy, SourceFile } from '../src/migration-strategy';
-
-const TEST_TIMEOUT = 500000;
 
 describe('migration-strategy', () => {
   describe('package', () => {
@@ -17,7 +15,7 @@ describe('migration-strategy', () => {
       const files: Array<SourceFile> = strategy.getMigrationOrder();
       const relativePaths: Array<string> = files.map((f) => f.relativePath);
       expect(relativePaths).toStrictEqual(['lib/a.js', 'index.js']);
-      expect(strategy.sourceType).toBe(DetectedSource.Library);
+      expect(strategy.sourceType).toBe(SourceType.Library);
     });
     test('simple with entrypoint', () => {
       const rootDir = getLibraryWithEntrypoint();
@@ -25,7 +23,7 @@ describe('migration-strategy', () => {
       const files: Array<SourceFile> = strategy.getMigrationOrder();
       const relativePaths: Array<string> = files.map((f) => f.relativePath);
       expect(relativePaths).toStrictEqual(['foo.js', 'depends-on-foo.js']);
-      expect(strategy.sourceType).toBe(DetectedSource.Library);
+      expect(strategy.sourceType).toBe(SourceType.Library);
     });
   });
 
@@ -36,75 +34,59 @@ describe('migration-strategy', () => {
       'app/components/salutation.js',
       'app/router.js',
     ];
-    test(
-      'app should match migration order',
-      async () => {
-        const project = await getEmberProjectFixture('app');
+    test('app should match migration order', async () => {
+      const project = await getEmberProjectFixture('app');
 
-        const strategy = getMigrationStrategy(project.baseDir);
-        const files: Array<SourceFile> = strategy.getMigrationOrder();
-        const actual: Array<string> = files.map((f) => f.relativePath);
-        expect(actual).toStrictEqual(EXPECTED_APP_FILES);
-        expect(strategy.sourceType).toBe(DetectedSource.EmberApp);
-      },
-      TEST_TIMEOUT
-    );
+      const strategy = getMigrationStrategy(project.baseDir);
+      const files: Array<SourceFile> = strategy.getMigrationOrder();
+      const actual: Array<string> = files.map((f) => f.relativePath);
+      expect(actual).toStrictEqual(EXPECTED_APP_FILES);
+      expect(strategy.sourceType).toBe(SourceType.EmberApp);
+    });
 
-    test(
-      'app-with-in-repo-addon should match migration order',
-      async () => {
-        const project = await getEmberProjectFixture('app-with-in-repo-addon');
+    test('app-with-in-repo-addon should match migration order', async () => {
+      const project = await getEmberProjectFixture('app-with-in-repo-addon');
 
-        const strategy = getMigrationStrategy(project.baseDir);
-        const files: Array<SourceFile> = strategy.getMigrationOrder();
-        const actual: Array<string> = files.map((f) => f.relativePath);
-        expect(actual).toStrictEqual([
-          'lib/some-addon/addon/components/greet.js',
-          'lib/some-addon/app/components/greet.js',
-          'lib/some-addon/index.js',
-          ...EXPECTED_APP_FILES,
-        ]);
-        expect(strategy.sourceType).toBe(DetectedSource.EmberApp);
-      },
-      TEST_TIMEOUT
-    );
+      const strategy = getMigrationStrategy(project.baseDir);
+      const files: Array<SourceFile> = strategy.getMigrationOrder();
+      const actual: Array<string> = files.map((f) => f.relativePath);
+      expect(actual).toStrictEqual([
+        'lib/some-addon/addon/components/greet.js',
+        'lib/some-addon/app/components/greet.js',
+        'lib/some-addon/index.js',
+        ...EXPECTED_APP_FILES,
+      ]);
+      expect(strategy.sourceType).toBe(SourceType.EmberApp);
+    });
 
-    test(
-      'app-with-in-repo-engine should match migration order',
-      async () => {
-        const project = await getEmberProjectFixture('app-with-in-repo-engine');
+    test('app-with-in-repo-engine should match migration order', async () => {
+      const project = await getEmberProjectFixture('app-with-in-repo-engine');
 
-        const strategy = getMigrationStrategy(project.baseDir);
-        const files: Array<SourceFile> = strategy.getMigrationOrder();
-        const actual: Array<string> = files.map((f) => f.relativePath);
-        expect(actual).toStrictEqual([
-          'lib/some-engine/addon/resolver.js',
-          'lib/some-engine/addon/engine.js',
-          'lib/some-engine/addon/routes.js',
-          'lib/some-engine/index.js',
-          ...EXPECTED_APP_FILES,
-        ]);
-        expect(strategy.sourceType).toBe(DetectedSource.EmberApp);
-      },
-      TEST_TIMEOUT
-    );
+      const strategy = getMigrationStrategy(project.baseDir);
+      const files: Array<SourceFile> = strategy.getMigrationOrder();
+      const actual: Array<string> = files.map((f) => f.relativePath);
+      expect(actual).toStrictEqual([
+        'lib/some-engine/addon/resolver.js',
+        'lib/some-engine/addon/engine.js',
+        'lib/some-engine/addon/routes.js',
+        'lib/some-engine/index.js',
+        ...EXPECTED_APP_FILES,
+      ]);
+      expect(strategy.sourceType).toBe(SourceType.EmberApp);
+    });
 
-    test(
-      'addon should match migration order',
-      async () => {
-        const project = await getEmberProjectFixture('addon');
+    test('addon should match migration order', async () => {
+      const project = await getEmberProjectFixture('addon');
 
-        const strategy = getMigrationStrategy(project.baseDir);
-        const files: Array<SourceFile> = strategy.getMigrationOrder();
-        const actual: Array<string> = files.map((f) => f.relativePath);
-        expect(actual).toStrictEqual([
-          'addon/components/greet.js',
-          'app/components/greet.js',
-          'index.js',
-        ]);
-        expect(strategy.sourceType).toBe(DetectedSource.EmberAddon);
-      },
-      TEST_TIMEOUT
-    );
+      const strategy = getMigrationStrategy(project.baseDir);
+      const files: Array<SourceFile> = strategy.getMigrationOrder();
+      const actual: Array<string> = files.map((f) => f.relativePath);
+      expect(actual).toStrictEqual([
+        'addon/components/greet.js',
+        'app/components/greet.js',
+        'index.js',
+      ]);
+      expect(strategy.sourceType).toBe(SourceType.EmberAddon);
+    });
   });
 });
