+++ packages/migration-graph/vite.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'vitest/config';
 //eslint-disable-next-line import/no-default-export
 export default defineConfig({
   test: {
-    testTimeout: 50_000,
+    testTimeout: 500_000,
     hookTimeout: 50_000,
     watchExclude: ['package.json', '**/fixtures/**'],
     threads: false,
